+++ packages/dartway_studio_bridge/lib/src/transport/client/studio_frame_controller_web.dart
@@ -3,17 +3,22 @@ import 'dart:ui_web' as ui_web;
 import 'package:web/web.dart' as web;
 
 import '../studio_message_channel.dart';
+import '../studio_message_drop.dart';
 import 'studio_client_web_channel.dart';
 import 'studio_frame_controller.dart';
 
 int _instanceCounter = 0;
 
-StudioFrameController createStudioFrameController({required String appUrl}) =>
-    _StudioFrameControllerWeb(appUrl);
+StudioFrameController createStudioFrameController({
+  required String appUrl,
+  StudioMessageDropObserver? onMessageDropped,
+}) => _StudioFrameControllerWeb(appUrl, onMessageDropped);
 
 class _StudioFrameControllerWeb implements StudioFrameController {
-  _StudioFrameControllerWeb(String appUrl)
-    : viewType = 'dartway-studio-frame-${_instanceCounter++}' {
+  _StudioFrameControllerWeb(
+    String appUrl,
+    StudioMessageDropObserver? onMessageDropped,
+  ) : viewType = 'dartway-studio-frame-${_instanceCounter++}' {
     _frame = web.document.createElement('iframe') as web.HTMLIFrameElement
       ..src = appUrl;
     _frame.style
@@ -24,7 +29,11 @@ class _StudioFrameControllerWeb implements StudioFrameController {
       viewType,
       (int viewId) => _frame,
     );
-    _channel = StudioClientWebChannel(_frame, Uri.parse(appUrl).origin);
+    _channel = StudioClientWebChannel(
+      _frame,
+      Uri.parse(appUrl).origin,
+      onMessageDropped: onMessageDropped,
+    );
   }
 
   late final web.HTMLIFrameElement _frame;